requirements.txt
Django==3.2.23
djangorestframework==3.13.1
python-decouple==3.6
psycopg2==2.9.3
drf-spectacular==0.27.0
Pillow==10.1.0
celery==5.3.6
